+++ RustyBags/src/SE_Bag.cs
@@ -1,5 +1,7 @@
 ﻿
 using System.Text;
+using BepInEx.Configuration;
+using RustyBags.Managers;
 using RustyBags.Utilities;
 using UnityEngine;
 
@@ -12,17 +14,30 @@ public class SE_Bag : SE_Stats
     public BagSetup data = null!;
     public int m_quality = 1;
     public float m_baseCarryWeight;
+    public float m_carryWeightPerQuality = 10f;
     public float m_inventoryWeightModifier;
+    
+    private ConfigEntry<float>? baseCarryWeightCfg;
+    private ConfigEntry<float>? inventoryWeightModifierCfg;
+    private float baseCarryWeight => baseCarryWeightCfg?.Value ?? m_baseCarryWeight;
+    private float inventoryWeightModifier => inventoryWeightModifierCfg?.Value ?? m_inventoryWeightModifier;
 
     public override void SetLevel(int itemLevel, float skillLevel)
     {
         m_quality = itemLevel;
-        if (m_baseCarryWeight > 0f) m_addMaxCarryWeight = m_baseCarryWeight + m_quality * 10;
+        if (baseCarryWeight > 0f) m_addMaxCarryWeight = baseCarryWeight + m_quality * m_carryWeightPerQuality;
     }
 
     public void ModifyInventoryWeight(ref float weight)
     {
-        weight *= Mathf.Max(1f - m_inventoryWeightModifier, 0f);
+        weight *= Mathf.Max(1f - inventoryWeightModifier, 0f);
+    }
+
+    public void SetupConfigs()
+    {
+        baseCarryWeightCfg = Configs.config(data.englishName, "Base Carry Weight", m_baseCarryWeight, "Setup base carry weight, increase by 10 per item quality");
+        inventoryWeightModifierCfg = Configs.config(data.englishName, "Inventory Weight Modifier", m_inventoryWeightModifier,
+            new ConfigDescription("Setup inventory weight multiplier", new AcceptableValueRange<float>(0f, 1f)));
     }
 
     public override string GetTooltipString()
@@ -32,7 +47,7 @@ public class SE_Bag : SE_Stats
         sb.Append(tooltip);
         if (m_inventoryWeightModifier != 0f)
         {
-            sb.AppendFormat("{0}: <color=orange>{1:+0;-0}%</color>\n", Keys.BagWeight, (m_inventoryWeightModifier - 1f) * 100);
+            sb.AppendFormat("{0}: <color=orange>{1:+0;-0}%</color>\n", Keys.BagWeight, (inventoryWeightModifier - 1f) * 100);
         }
         if (data.sizes.TryGetValue(m_quality, out var size))
         {