+++ RustyBags/Thunderstore/CHANGELOG.md
@@ -1,2 +1,15 @@
+# 1.0.3
+- Added `Extended Player Inventory` API
+- If `EPI API` is installed, new slot is added: `Bag`
+- Added Crossbow Bolt Quiver
+- Added carry weight config
+- Added restriction config
+
+# 1.0.2
+- Updated bepinex dependency string to 5.4.2333
+
+# 1.0.1
+- Added YmlDotNet as dependency string
+
 # 1.0.0
 - Initial release